lib/pyproject.toml
[build-system]
requires = ["setuptools>=45", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "libregistry"
version = "0.1.0"
description = "A library to manage system configuration like a registry"
authors = [
    {name = "MVDW-Java"},
]
requires-python = ">=3.6"
dependencies = [
    "tomli>=1.1.0;python_version<'3.11'",
]
readme = "README.md"
license = {text = "GPL-3.0-or-later"}

[tool.setuptools.packages.find]
where = ["."]
include = ["libregistry*"]
